everything works on pytorch==0.4

+++ arc_solvers/nn/util.py
@@ -28,7 +28,7 @@ def masked_mean(tensor, dim, mask):
     # count
     count_tensor = torch.sum((mask != 0), dim)
     # set zero count to 1 to avoid nans
-    zero_count_mask = (count_tensor == 0)
+    zero_count_mask = (count_tensor == 0).long()
     count_plus_zeros = (count_tensor + zero_count_mask).float()
     # average
     mean_tensor = total_tensor / count_plus_zeros
